refactor(trials): align trial pdf code with web conventions
- move trial PDF helpers into the feature-scoped `apps/web/lib/public/beagle/trials/pdf` tree
- remove unused `@contracts` aliases, simplify the admin PDF link, and normalize the PDF renderer structure

Refs: BEJ-89
