+++ apps/web/lib/public/beagle/trials/index.ts
@@ -1,5 +1,6 @@
 export * from "./clipboard";
 export * from "./date";
 export * from "./search-request";
+export * from "./trial-pdf-route";
 export * from "./trial-route";
 export * from "./types";
