+++ apps/web/lib/public/beagle/trials/trial-pdf-route.ts
