--- apps/web/lib/trials/index.ts
@@ -1,1 +1,0 @@
-export { getTrialPdfHref } from "./trial-pdf-route";
