+++ packages/auth/tsconfig.json
@@ -3,8 +3,6 @@
   "compilerOptions": {
     "paths": {
       "@auth/*": ["./*"],
-      "@contracts": ["../contracts/index.ts"],
-      "@contracts/*": ["../contracts/*"],
       "@db/*": ["../db/*"]
     }
   },
